\name{extractoR.content-package}
\alias{extractoR.content-package}
\alias{extractoR.content}
\docType{package}
\encoding{UTF-8}
\title{CRAN packages metadata extractor (content)}
\description{
  extractoR.extract is part of extractoR which is a set of R packages
  that can be used to fetch, extract and query CRAN packages metadata.
  extractoR.content browses packages and aggregates informations about
  packages content.
}

\author{
Maëlick Claes <himself@maelick.net>

Maintainer: Maëlick Claes <himself@maelick.net>
}

\examples{
\dontrun{
library(extractoR.content)
}
}